+++ packages/components/button/src/button.tsx
@@ -1,4 +1,4 @@
-import type { HTMLUIProps, ThemeProps, CSSUIObject } from "@yamada-ui/core"
+import type { HTMLUIProps, ThemeProps, CSSUIObject, FC } from "@yamada-ui/core"
 import {
   ui,
   forwardRef,
@@ -9,7 +9,7 @@ import type { LoadingProps } from "@yamada-ui/loading"
 import { Loading as LoadingIcon } from "@yamada-ui/loading"
 import { Ripple, useRipple } from "@yamada-ui/ripple"
 import { cx, merge, dataAttr, mergeRefs } from "@yamada-ui/utils"
-import type { ElementType, FC, ReactElement } from "react"
+import type { ElementType, ReactElement } from "react"
 import { useCallback, useMemo, useRef } from "react"
 import { useButtonGroup } from "./button-group"
 
@@ -201,6 +201,9 @@ export const Button = forwardRef<ButtonProps, "button">(
   },
 )
 
+Button.displayName = "Button"
+Button.__ui__ = "Button"
+
 const Loading: FC<
   Pick<ButtonProps, "className" | "loadingIcon" | "loadingText">
 > = ({ className, loadingIcon, loadingText }) => {
@@ -230,6 +233,9 @@ const Loading: FC<
   )
 }
 
+Loading.displayName = "Loading"
+Loading.__ui__ = "Loading"
+
 const Content: FC<
   Pick<
     ButtonProps,
@@ -238,6 +244,7 @@ const Content: FC<
 > = ({ startIcon, leftIcon, endIcon, rightIcon, children }) => {
   startIcon ??= leftIcon
   endIcon ??= rightIcon
+
   return (
     <>
       {startIcon ? <Icon>{startIcon}</Icon> : null}
@@ -246,6 +253,9 @@ const Content: FC<
     </>
   )
 }
+
+Content.displayName = "Content"
+Content.__ui__ = "Content"
 
 const Icon: FC<HTMLUIProps<"span">> = ({ children, className, ...rest }) => {
   return (
@@ -261,6 +271,9 @@ const Icon: FC<HTMLUIProps<"span">> = ({ children, className, ...rest }) => {
     </ui.span>
   )
 }
+
+Icon.displayName = "Icon"
+Icon.__ui__ = "Icon"
 
 export const useButtonType = (value?: ElementType) => {
   const isButton = useRef(!value)