+++ packages/components/button/src/icon-button.tsx
@@ -48,3 +48,6 @@ export const IconButton = forwardRef<IconButtonProps, "button">(
     )
   },
 )
+
+IconButton.displayName = "IconButton"
+IconButton.__ui__ = "IconButton"