+++ packages/components/modal/src/modal-close-button.tsx
@@ -29,3 +29,6 @@ export const ModalCloseButton = forwardRef<ModalCloseButtonProps, "button">(
     )
   },
 )
+
+ModalCloseButton.displayName = "ModalCloseButton"
+ModalCloseButton.__ui__ = "ModalCloseButton"