+++ packages/components/modal/src/modal-overlay.tsx
@@ -40,3 +40,6 @@ export const ModalOverlay = motionForwardRef<ModalOverlayProps, "div">(
     )
   },
 )
+
+ModalOverlay.displayName = "ModalOverlay"
+ModalOverlay.__ui__ = "ModalOverlay"