+++ packages/components/progress/src/circle-progress.tsx
@@ -11,8 +11,7 @@ import {
   omitThemeProps,
 } from "@yamada-ui/core"
 import { useAnimation } from "@yamada-ui/use-animation"
-import { useToken } from "@yamada-ui/use-token"
-import { cx, isUnit, replaceObject, valueToPercent } from "@yamada-ui/utils"
+import { cx, valueToPercent } from "@yamada-ui/utils"
 import type { FC } from "react"
 
 interface CircleProgressOptions {
@@ -88,7 +87,7 @@ export const CircleProgress = forwardRef<CircleProgressProps, "div">(
     let {
       className,
       children,
-      boxSize = "6em",
+      boxSize = "6rem",
       thickness = "0.625rem",
       color = "primary",
       trackColor = "border",
@@ -100,10 +99,6 @@ export const CircleProgress = forwardRef<CircleProgressProps, "div">(
       speed = ["1.4s", "2s"],
       ...rest
     } = omitThemeProps(mergedProps)
-
-    thickness = replaceObject(thickness, (value) =>
-      !isUnit(value) ? useToken("sizes", value) : value,
-    )
 
     const isTransparent = value === 0 && !isAnimation
     const percent = valueToPercent(value, min, max)
@@ -132,7 +127,11 @@ export const CircleProgress = forwardRef<CircleProgressProps, "div">(
 
     const css: CSSUIObject = {
       ...styles,
-      fontSize: boxSize,
+      vars: [
+        { name: "boxSize", token: "sizes", value: boxSize },
+        { name: "thickness", token: "sizes", value: thickness },
+      ],
+      fontSize: "$boxSize",
     }
 
     const circleProps: CircleProps = isAnimation
@@ -156,10 +155,10 @@ export const CircleProgress = forwardRef<CircleProgressProps, "div">(
         {...rest}
       >
         <Shape boxSize={boxSize} isAnimation={isAnimation} speed={speed}>
-          <Circle stroke={trackColor} strokeWidth={thickness} />
+          <Circle stroke={trackColor} strokeWidth="$thickness" />
           <Circle
             stroke={color}
-            strokeWidth={thickness}
+            strokeWidth="$thickness"
             strokeLinecap={isRounded ? "round" : undefined}
             opacity={isTransparent ? 0 : undefined}
             {...circleProps}