+++ packages/components/select/src/select-icon.tsx
@@ -1,9 +1,13 @@
-import type { CSSUIObject, HTMLUIProps } from "@yamada-ui/core"
+import type {
+  CSSUIObject,
+  FunctionComponent,
+  HTMLUIProps,
+} from "@yamada-ui/core"
 import { ui, forwardRef } from "@yamada-ui/core"
 import { ChevronIcon, CloseIcon } from "@yamada-ui/icon"
 import { useClickable } from "@yamada-ui/use-clickable"
 import { cx, getValidChildren, isValidElement } from "@yamada-ui/utils"
-import type { FC, HTMLAttributes } from "react"
+import type { HTMLAttributes } from "react"
 import { cloneElement, useRef } from "react"
 import { useSelectContext } from "./use-select"
 
@@ -45,11 +49,14 @@ export const SelectIcon = forwardRef<SelectIconProps, "div">(
   },
 )
 
+SelectIcon.displayName = "SelectIcon"
+SelectIcon.__ui__ = "SelectIcon"
+
 export interface SelectClearIconProps extends SelectIconProps {
   disabled?: boolean
 }
 
-export const SelectClearIcon: FC<SelectClearIconProps> = ({
+export const SelectClearIcon: FunctionComponent<SelectClearIconProps> = ({
   className,
   children,
   ...props
@@ -76,3 +83,6 @@ export const SelectClearIcon: FC<SelectClearIconProps> = ({
     </SelectIcon>
   )
 }
+
+SelectClearIcon.displayName = "SelectClearIcon"
+SelectClearIcon.__ui__ = "SelectClearIcon"