+++ packages/components/stat/src/stat.tsx
@@ -128,3 +128,5 @@ export const Stat = forwardRef<StatProps, "dl">((props, ref) => {
     </StatProvider>
   )
 })
+Stat.displayName = "Stat"
+Stat.__ui__ = "Stat"