+++ packages/components/table/src/thead.tsx
@@ -1,3 +1,4 @@
+import { FC } from "@yamada-ui/core"
 import type { CSSUIObject } from "@yamada-ui/core"
 import type { IconProps } from "@yamada-ui/icon"
 import { Icon, ChevronIcon } from "@yamada-ui/icon"
@@ -12,14 +13,13 @@ import {
   useTableStyles,
 } from "@yamada-ui/native-table"
 import { runIfFunc, handlerAll, cx } from "@yamada-ui/utils"
-import type { FC } from "react"
 import { useMemo } from "react"
 import type { SortDirection, Column } from "./use-table"
 import { useTableContext, render } from "./use-table"
 
 export interface TableHeadProps extends NativeTableHeadProps {}
 
-export const Thead = ({ ...rest }: TableHeadProps) => {
+export const Thead: FC = ({ ...rest }: TableHeadProps) => {
   const { headerGroups, headerGroupProps, headerProps, sortIconProps } =
     useTableContext()
 
@@ -131,6 +131,9 @@ export const Thead = ({ ...rest }: TableHeadProps) => {
     </NativeThead>
   )
 }
+
+Thead.displayName = "Thead"
+Thead.__ui__ = "Thead"
 
 export interface SortIconProps extends IconProps {
   isSorted: false | SortDirection